requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pydub>=0.25.1
gunicorn>=21.2.0
python-dotenv>=1.0.0
